changing int type to float type

+++ Particles.py
@@ -6,9 +6,9 @@ class Particles:
         self.masses = masses
     def size(self) -> int:
         return len(self.positions)
-    def at(self, index : int) -> ((int, int, int), int, int): #position, charge, mass
+    def at(self, index : int) -> ((float, float, float), float, float): #position, charge, mass
         return (self.positions[index], self.charges[index], self.masses[index])
-    def update(self, index: int, x: int, y: int, z: int) -> bool:
+    def update(self, index: int, x: float, y: float, z: float) -> bool:
         if(index >= len(self.positions)):
             return False
         self.positions[index] = numpy.asarray([x, y, z])
